Fixed problem if ship being build at the base have 0 for beamId or launcherId

+++ src/main/scala/replanets/common/BeamspecItem.scala
@@ -9,7 +9,7 @@ case class BeamspecItem(
   name: String,
   cost: Cost,
   mass: Short,
-  techLevel: Short,
+  techLevel: Int,
   killValue: Short,
   damageValue: Short
 )
